Consistently use qint64 for session time (#17)
* Consistently use qint64 for session time

The problem is that QDateTime::currentMSecsSinceEpoch returns qint64.
Comparing timestamps then leads to comparisons with unsigned.
HttpSession::getLastAccess returns qint64.

This fixes a compilation warning about comparing signed/unsigned with clang:
qtwebapp/QtWebApp/httpserver/httpsessionstore.cpp:105:27:
warning: comparison of integers of different signs: 'long long' and 'quint64' (aka 'unsigned long long') [-Wsign-compare]
        if (now-lastAccess>cfg.expirationTime)
            ~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~

* Fix formatting string for expiration time

The variable is 64 bit, so use lli. Allows compiling with clang with -Werror.
QtWebApp/httpserver/httpsessionstore.cpp:19:70: error: format specifies type 'int' but the argument has type 'qint64' (aka 'long long') [-Werror,-Wformat]
    qDebug("HttpSessionStore: Sessions expire after %i milliseconds",cfg.expirationTime);
                                                    ~~               ^~~~~~~~~~~~~~~~~~
                                                    %lli

+++ QtWebApp/httpserver/httpserverconfig.cpp
@@ -57,7 +57,7 @@ HttpSessionStoreConfig::HttpSessionStoreConfig(QSettings *settings)
 
 void HttpSessionStoreConfig::parseSettings(const QSettings &settings)
 {
-	expirationTime = parseNum(settings.value("expirationTime", (qulonglong)expirationTime), 1000);
+	expirationTime = parseNum(settings.value("expirationTime", expirationTime), 1000);
 	cookieName = settings.value("cookieName", cookieName).toByteArray();
 	
 	cookiePath = settings.value("cookiePath", cookiePath).toByteArray();
